--- cooee.db-config.json
@@ -1,62 +1,0 @@
-{
-    "export-query": "SELECT * FROM RepositoryObject",
-    "tables": {
-        "RepositoryObject": {
-            "all_props": [
-                "linguisticGenre",
-                "communicationMode",
-                "citation.materialType",
-                "citation.name",
-                "citation.pageEnd",
-                "@type",
-                "author",
-                "@id",
-                "citation",
-                "inLanguage",
-                "citation.wordCount",
-                "dateCreated",
-                "indexableText",
-                "name",
-                "citation.@id",
-                "citation.isPartOf",
-                "citation.pageStart",
-                "citation.@type",
-                "conformsTo",
-                "hasPart",
-                "register"
-            ],
-            "ignore_props": [
-                "communicationMode"
-            ],
-            "expand_props": [
-                "citation"
-            ]
-        },
-        "Person": {
-            "all_props": [
-                "linguisticGenre",
-                "communicationMode",
-                "birthPlace",
-                "@type",
-                "author",
-                "@id",
-                "class",
-                "inLanguage",
-                "citation",
-                "prov:specializationOf",
-                "dateCreated",
-                "indexableText",
-                "age",
-                "name",
-                "birthDate",
-                "person",
-                "immigration",
-                "conformsTo",
-                "hasPart",
-                "register"
-            ],
-            "ignore_props": [],
-            "expand_props": []
-        }
-    }
-}
