+++ ro_crate_to_sqlite/__init__.py
@@ -1,0 +1,3 @@
+from .index import build, add_csv, flatten_entities
+
+__all__ = ['build', 'add_csv', 'flatten_entities']
